Added tomato and lettuce crop loot table
