+++ src/main/java/net/moritz_htk/bettermcdonaldsmod/data/BMMLootTableProvider.java
@@ -4,6 +4,7 @@ package net.moritz_htk.bettermcdonaldsmod.data;
 import net.minecraft.data.PackOutput;
 import net.minecraft.data.loot.LootTableProvider;
 import net.minecraft.world.level.storage.loot.parameters.LootContextParamSets;
+import net.moritz_htk.bettermcdonaldsmod.loot.BMMBlockLootTables;
 
 import java.util.List;
 import java.util.Set;
